cleaned up the layout
