+++ Cargo.toml
@@ -67,6 +67,7 @@ tiff = "0.11.3"
 toml = "0.9.8"
 uuid = { version = "1.22.0", features = ["v5"] }
 walkdir = "2.5.0"
+dmg_layout = "0.1.0"
 
 [dev-dependencies]
 winit = "0.30.13"
